Route info and route gen fix

+++ app/src/main/java/com/example/pbrg_android/data/model/HoldData.kt
@@ -4,6 +4,6 @@ package com.example.pbrg_android.data.model
  * Data class that encodes the x and y coordinates of a hold
  */
 data class HoldData(
-    val x: Float,
-    val y: Float
+    val x: Double,
+    val y: Double
 )
